Moves test directory to root
